+++ modules/tinymce/src/themes/silver/main/ts/modes/Iframe.ts
@@ -155,13 +155,10 @@ const render = (editor: Editor, uiComponents: RenderUiComponents, rawUiConfig: R
   }
 
   const api: Partial<EditorUiApi> = {
-    enable: () => {
-      ReadOnly.broadcastReadonly(uiComponents, false);
+    setEnabled: (state) => {
+      ReadOnly.broadcastReadonly(uiComponents, !state);
     },
-    disable: () => {
-      ReadOnly.broadcastReadonly(uiComponents, true);
-    },
-    isDisabled: () => Disabling.isDisabled(outerContainer)
+    isEnabled: () => !Disabling.isDisabled(outerContainer)
   };
 
   return {