+++ modules/tinymce/src/themes/silver/test/ts/browser/editor/DisableTest.ts
@@ -32,24 +32,24 @@ describe('browser.tinymce.themes.silver.editor.DisableTest', () => {
       })),
       TinyDom.container(editor)
     );
-    assert.equal(editor.ui.isDisabled(), disabled, 'Editor isDisabled should return current disabled state');
+    assert.equal(!editor.ui.isEnabled(), disabled, 'Editor isEnabled should return current disabled state');
   };
 
   context('Test disable/enable APIs', () => {
     it('TINY-6397: Should be able to enable and disable the UI', async () => {
       const editor = hook.editor();
-      editor.ui.disable();
+      editor.ui.setEnabled(false);
       await pAssertUiDisabled(editor, true);
-      editor.ui.enable();
+      editor.ui.setEnabled(true);
       await pAssertUiDisabled(editor, false);
     });
 
     it('TINY-6397: Should not be able to enable the UI when in readonly mode', async () => {
       const editor = hook.editor();
-      editor.ui.disable();
+      editor.ui.setEnabled(false);
       editor.mode.set('readonly');
       await pAssertUiDisabled(editor, true);
-      editor.ui.enable();
+      editor.ui.setEnabled(true);
       await pAssertUiDisabled(editor, true);
       editor.mode.set('design');
     });